requirements.txt
# JARVIS 1.5 — Python Dependencies
# SYSTEMS™ · architectofscale.com

# ── Core ──────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0

# ── Database (Supabase + Postgres) ───────────────
supabase>=2.3.0
httpx>=0.27.0
asyncpg>=0.30.0
redis>=5.0.0

# ── LLM Integration ─────────────────────────────
litellm>=1.50.0
openai>=1.50.0

# ── Telegram Bot ─────────────────────────────────
python-telegram-bot>=21.0

# ── Web API ──────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# ── Task Scheduling ─────────────────────────────
apscheduler>=3.10.0

# ── Logging & Monitoring ────────────────────────
structlog>=24.1.0

# ── Voice (TTS/STT) ────────────────────────────
edge-tts>=6.1.0

# ── File Processing ───────────────────────────
pypdf>=4.0.0

# ── Utilities ────────────────────────────────────
aiohttp>=3.10.0
tenacity>=9.0.0
